+++ src/Profile.php
@@ -25,16 +25,16 @@ use function trim;
 final class Profile implements JsonSerializable
 {
     /** @var non-empty-string */
-    private readonly string $label;
-    private ?Snapshot $start = null;
-    private ?Snapshot $end = null;
-    private Metrics $metrics;
+    public readonly string $label;
+    public readonly Snapshot $start;
+    public readonly Snapshot $end;
+    public readonly Metrics $metrics;
 
     /**
      * @throws InvalidArgument
      * @throws RandomException
      */
-    public function __construct(?string $label = null)
+    public function __construct(?string $label, Snapshot $start, Snapshot $end)
     {
         $label ??= self::randomLabel();
         $label = strtolower(trim($label));
@@ -45,7 +45,9 @@ final class Profile implements JsonSerializable
         1 === preg_match('/^[a-z0-9][a-z0-9_]*$/', $label) || throw new InvalidArgument('The label must start with a lowercased letter or a digit and only contain lowercased letters, digits, or underscores.');
 
         $this->label = $label;
-        $this->metrics = Metrics::none();
+        $this->start = $start;
+        $this->end = $end;
+        $this->metrics = Metrics::fromSnapshots($start, $end);
     }
 
     /**
@@ -56,14 +58,6 @@ final class Profile implements JsonSerializable
     public static function randomLabel(): string
     {
         return bin2hex(random_bytes(6));
-    }
-
-    /**
-     * @return non-empty-string
-     */
-    public function label(): string
-    {
-        return $this->label;
     }
 
     /**
@@ -85,68 +79,5 @@ final class Profile implements JsonSerializable
     public function jsonSerialize(): array
     {
         return $this->stats();
-    }
-
-    public function hasNotBegun(): bool
-    {
-        return null === $this->start;
-    }
-
-    public function hasBegun(): bool
-    {
-        return null !== $this->start;
-    }
-
-    public function isRunning(): bool
-    {
-        return null !== $this->start
-            && null === $this->end;
-    }
-
-    public function hasEnded(): bool
-    {
-        return null !== $this->start
-            && null !== $this->end;
-    }
-
-    public function beginProfiling(): void
-    {
-        (null === $this->start && null === $this->end) || throw new InvalidProfileState('Profiling cannot be started if it has already started.');
-
-        $this->start = Snapshot::now();
-    }
-
-    public function endProfiling(): void
-    {
-        (null !== $this->start && null === $this->end) || throw new InvalidProfileState('Profiling cannot be ended if it is not running.');
-
-        $this->end = Snapshot::now();
-        $this->metrics = Metrics::fromSnapshots($this->start, $this->end);
-    }
-
-    /**
-     * @throws InvalidProfileState
-     */
-    public function start(): Snapshot
-    {
-        return $this->start ?? throw new InvalidProfileState('Profiling has yet to be started.');
-    }
-
-    /**
-     * @throws InvalidProfileState
-     */
-    public function end(): Snapshot
-    {
-        return $this->end ?? throw new InvalidProfileState('Profiling has yet to be ended.');
-    }
-
-    /**
-     * @throws InvalidProfileState
-     */
-    public function metrics(): Metrics
-    {
-        $this->hasEnded() || throw new InvalidProfileState('Profiling has yet to be ended.');
-
-        return $this->metrics;
     }
 }