+++ src/ProfilingResult.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Bakame\Aide\Profiler;
 
+use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
+
 final class ProfilingResult
 {
     public function __construct(
@@ -12,13 +15,22 @@ final class ProfilingResult
     ) {
     }
 
-    public static function profile(?string $label, callable $callback, mixed ...$args): self
-    {
-        $profile = new Profile($label);
-        $profile->beginProfiling();
-        $result = ($callback)(...$args);
-        $profile->endProfiling();
+    /**
+     * @throws \Random\RandomException
+     */
+    public static function profile(
+        string $label,
+        callable $callback,
+        LoggerInterface $logger = new NullLogger(),
+        mixed ...$args
+    ): self {
+        $logger->info('Starting profiling for label: '.$label.'.');
+        $start = Snapshot::now();
+        $value = ($callback)(...$args);
+        $end = Snapshot::now();
+        $profile = new Profile($label, $start, $end);
+        $logger->info('Finished profiling for label: '.$label.'.', $profile->stats());
 
-        return new self($profile, $result);
+        return new self($profile, $value);
     }
 }